oc_projet9/requirements.txt
asgiref==3.7.2
astroid==2.15.6
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.1.0
click==8.1.7
coverage==7.3.1
dill==0.3.7
Django==4.2.3
django-bootstrap-icons==0.8.3
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
exceptiongroup==1.1.3
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
isort==5.12.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.10.0
pluggy==1.3.0
psycopg==3.1.10
psycopg-binary==3.1.10
pycodestyle==2.11.0
pyflakes==3.1.0
Pygments==2.15.1
pytest==7.4.2
pytest-django==4.5.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.6.3
urllib3==2.0.3
wrapt==1.15.0
